Implicit conversion fixes

+++ clock.cpp
@@ -75,7 +75,7 @@ namespace octo
     void VariableClock::onTick()
     {
         auto now = std::chrono::high_resolution_clock::now();
-        rate_ = 1.0 / std::chrono::duration_cast<std::chrono::duration<double>>(now - lastNow).count();
+        rate_ = 1.0f / std::chrono::duration_cast<std::chrono::duration<float>>(now - lastNow).count();
         lastNow = now;
         
         ++timestamp;
